# back-end/documentação/swagger.yaml
swagger: '2.0'
info:
  description: "Documentação da API do projeto. Para editar esse arquivo é recomendado utilizar a extensão do VSCode 'Swagger Viewer'"
  version: '1.0.0'
  title: 'Projeto'
host: 'https://task.compjunior.com.br/'
basePath: '/api'
...
schemes:
  - 'https'
paths:
  /task:
    get:
      tags:
        - 'task'
      summary: 'RESUMO EXEMPLO GET'
      description: 'DESCRIÇÃO EXEMPLO GET'
      operationId: 'login'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      responses:
        '200':
          description: 'Deu tudo certo'
          schema:
            type: 'array'
            items:
              type: 'object'
              properties:
                dado:
                  type: 'string'
                  task: 'task'
                dado2:
                  type: 'string'
                  task: 'task2'
        '400':
          description: 'Bad request'
        '500':
          description: 'Internal server error'
    post:
      tags:
        - 'task'
      summary: 'RESUMO EXEMPLO POST'
      description: 'DESCRIÇÃO EXEMPLO POST'
      operationId: 'signUp'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Dados de cadastro do usuário'
          required: true
          schema:
            type: 'object'
            properties:
              dado:
                type: 'string'
                task: 'task'
              dado2:
                type: 'string'
                task: 'task2'
      responses:
        '200':
          description: 'Deu tudo certo'
        '400':
          description: 'Bad request'
        '500':
          description: 'Internal server error'
...
securityDefinitions:
  petstore_auth:
    type: 'oauth2'
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: 'implicit'
    scopes:
      write:palestrantes: 'modify pets in your account'
  api_key:
    type: 'apiKey'
    name: 'api_key'
    in: 'header'
...